Autonmus Coding Agent/requirements.txt
# pyproject.toml
[tool.poetry.dependencies]
python = "^3.10"
google-generativeai = "^0.3.0"
chromadb = "^0.4.0"  # Vector search
tiktoken = "^0.5.0"  # Token counting
rich = "^13.0.0"  # Beautiful CLI output
pydantic = "^2.0.0"  # Data validation
fastapi = "^0.109.0"  # For future API
redis = "^5.0.0"  # State management